improve: organise code in Clare.java
move commands to a different package, add comments for public methods
