+++ data/data.txt
@@ -1,2 +1,4 @@
 T|0|sas
 D|0|hihi|2025-08-28 14:34:55.688743
+T|0|
+D|0|12|2025-08-28 15:07:58.528065800
