+++ src/main/java/commands/Commands.java
@@ -1,0 +1,13 @@
+package commands;
+
+public enum Commands {
+    LIST,
+    BYE,
+    MARK,
+    UNMARK,
+    DELETE,
+    TODO,
+    DEADLINE,
+    EVENT,
+    FIND,
+}
